Varnish: cache mobile and non-mobile version separately
Until now, it was possible that a desktop user got a mobile version of a page and vice versa. Using User-Agent detection, this should fix that.

Using the 'toggle buttons' to show a mobile version of a page to a desktop user and vice versa (showing a desktop version to a mobile user) still doesn't work really well, but this is more important.