move ndkilla from ops to mw-roots + cache-admins
Effective immediately, I don't really trust myself with access related to non-mediawiki related things. I understand that this was kind of an odd situation related to the fact that nobody on the ops team handled a critical alert for over 14 days, but I still don't blame anybody besides myself. There were several things I did that resulted in downtime. In addition to a couple stupid things that started an avalanche, I didn't even know how to depool the one problematic server, and spent several minutes looking through puppet and mw-config and dns history trying to figure out how to depool an mw server.

I'd like to keep github admin/owner/whatever until this is discussed so that I may add the rights to myself in the event of an emergency, but I feel that this needs some serious thought, and as such I'm doing this until such thought and conversation happens.