Cross out database and user creation
For now doing this on the command like will fail as it needs the user and password. But i am not sure how to do this as on the command line it will do a password prompt, but that won't work if you are doing it through puppet.

Maybe we will have to in the future at some point create a user that dosen't require a password that is a super user?