T7319: Change KillMode to control-group for Varnish (#1857)
KillMode of process is not recommended as it silently kills the process but allows it to continue running unaware to system.d. In essence, this means an OOM killer would 'kill' a sub process, not trigger the master service to be killed and consequently keep running broken until restarted manually.