php: drop max_requests by 50% (#2168)
The previous mitigations have reduced memory by quite a bit and limited the rate of OOMs but the graphs still aren't amazing.

Restarting an individual worker isn't much of an impact for user facing performance but it might help flatten the graphs a bit more and limit the leak even further. This should help us increase confidence in bullseye even if we haven't found the root cause and limit any impact of any issue which is always good.