php: lower pm.max_requests to 1,000
I'm not sure this is needed yet but I'm creating just in case. I'll review the graphs over tomorrow.

I kinda hope not because our other changes should have had a big impact but we still need to see what the graph looks like over a longer period and what happens. I'm going to try and decide by tomorrow evening because it's getting close to Christmas.