Use python `write()` instead of `echo()`
Also tabs are invalid in yaml syntax. Use two spaces; remove punctuation from yaml key, to strip IE periods from it. This may actually work without doing it, but to maintain how we have previously been doing it, we strip it instead.

Also, we use `write()` instead of `writelines()` here, only because this looks nicer in a code standard, in my own opinion.