Use SRI when fetching Bootstrap CSS (#3124)
This PR modifies the HTML shown on "Backend Fetch Failed" errors to use
SRI when retrieving the Bootstrap CSS from jsdelivr. SubResource
Integrity is a security measure to be used when retrieving content on
CDNs to prevent a series of possible attacks. For an introduction on
SRI, see
https://developer.mozilla.org/en-US/docs/Web/Security/Subresource_Integrity.

The hash is a SHA-384 hash created by me using OpenSSL 3.0.7. I
generated it from both the Bootstrap GitHub 4.6.0 release
(https://github.com/twbs/bootstrap/releases/tag/v4.6.0) and the file
offered for download by jsdelivr, both returning the same hash.