mariadb remove read-only on startup
mysql should automatically handle error recovery. There's no sense making it read-only especially if we don't have a procedure in place to validate the database after it's started up.

The only use case I can think of for having read-only on startup is for doing maintenance where the server is expected to restart, but we don't want any changes after the restart.