all_certs.pp assume nginx service exists (#3396)
Puppet is currently not creating the proper notify relationship between
`/etc/ssl/localcerts` and `Service['nginx']`. I'm assuming this is
because `defined(Service['nginx'])` is returning false for some reason
(possibly evaluation order?). For now, we can always assume we can reach
it, given that all_certs is always called inside an nginx definition
file.

Maybe we need a requires?