MirahezeLogbot: Don't join #miraheze-sre-security
I previously tried this in #1757 and it was reverted in #1759 as only one made it not being able to join at all, but I realized why: if there is not a comma after a single-element tuple, Python will not recognize it as a tuple at all, amd thus the for loop that makes it connect to IRC channels does nothing. So I want to see if this works. If it doesn't, we can revert again.