-----BEGIN PGP SIGNED MESSAGE----- Hash: SHA512
This commit adds a new key for my shell account at WikiTide. It is also an ed25519-sk key, created with a newly purchased YubiKey. It is meant as a backup in case my main key breaks.

This commit message, as well as the commit itself, is signed with my OpenPGP key.
-----BEGIN PGP SIGNATURE-----

iHUEARYKAB0WIQQK+0J/GAD9iXUcQDUpIihzWucH/wUCZip0vgAKCRApIihzWucH
/wN4AQCEV5+hVEhNempO6+LHAOdJhcSoUeVJY9ua4Ntnjyme3wD/YtI+iVCIeX83
1SYvSTXAtd07EgQGxHvlftB/chk3gQ8=
=tx1s
-----END PGP SIGNATURE-----