cache::haproxy: modify log-format (#4359)
Put http status nearer to the front so it's easily findable using awk.
This is because of the http user agent getting in the way. And in
addition we also put response_size nearer to the front. This is only
done when logging to disk.

We also remove some unneeded logging properties.