changeprop201: Lower concurrency for LocalGlobalUserPageCacheUpdateJob to 25 (#4488)
Having it at 100 causes the cpu to jump to 100% on mwtask and also the
load gets very high. This is counter productive and won't make things go
any faster. It does the opposite and will make things slower, as well as
taking resources away from other jobs as well. Let's lower this to 25.